requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
